+++ internal/cli/upgrade.go
@@ -6,12 +6,12 @@ import (
 	"sort"
 	"sync"
 
-	chiselalpm "github.com/kodos-prj/chisel/pkg/alpm"
-	"github.com/kodos-prj/chisel/pkg/aur"
-	"github.com/kodos-prj/chisel/pkg/config"
-	"github.com/kodos-prj/chisel/pkg/download"
-	"github.com/kodos-prj/chisel/pkg/registry"
-	"github.com/kodos-prj/chisel/pkg/store"
+	chiselalpm "github.com/kodos-prj/pistacho/pkg/alpm"
+	"github.com/kodos-prj/pistacho/pkg/aur"
+	"github.com/kodos-prj/pistacho/pkg/config"
+	"github.com/kodos-prj/pistacho/pkg/download"
+	"github.com/kodos-prj/pistacho/pkg/registry"
+	"github.com/kodos-prj/pistacho/pkg/store"
 )
 
 // UpgradeCommand implements the 'chisel upgrade' command.